suchi_phase1_pack/scripts/nci-ingestion/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
rdflib>=6.3.0
pyyaml>=6.0
python-dotenv>=1.0.0
markdownify>=0.11.6
urllib3>=2.0.0














